Adding ResponseEntities to the RestController

+++ src/main/java/com/kemal/spring/web/controllers/restApiControllers/restController.java
@@ -2,6 +2,8 @@ package com.kemal.spring.web.controllers.restApiControllers;
 
 import com.kemal.spring.domain.User;
 import com.kemal.spring.service.UserService;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -22,12 +24,21 @@ public class restController {
     }
 
     @GetMapping("/adminPage/json-users")
-    public List<User> getUsers() {
-        return userService.findAll();
+    public ResponseEntity<List<User>> getUsers() {
+        List<User> allUsers = userService.findAll();
+        if(allUsers == null) return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        else if(allUsers.isEmpty()) return new ResponseEntity<>(allUsers, HttpStatus.NOT_FOUND);
+        else return new ResponseEntity<>(allUsers, HttpStatus.OK);
     }
 
     @PostMapping ("/adminPage/json-users/delete/{id}")
-    public void deleteUser (@PathVariable Long id){
-        userService.delete(id);
+    public ResponseEntity<User> deleteUser (@PathVariable Long id){
+        User userToDelete = userService.findById(id);
+        if(userToDelete == null) return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+
+        else {
+            userService.delete(id);
+            return new ResponseEntity<>(userToDelete, HttpStatus.NO_CONTENT);
+        }
     }
 }
